--- bootstrap/src/Facades/Filter.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Bootstrap\Facades;
-
-class Filter extends Facade
-{
-    protected static function getFacadeAccessor()
-    {
-        return 'filter';
-    }
-}
