+++ components/Facades/Slayer/Auth.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Bootstrap\Facades;
+namespace Components\Facades\Slayer;
+
+use Bootstrap\Facades\Facade;
 
 class Auth extends Facade
 {
